refactoring+cleanup
