+++ server/core/cmd/httpserver/handler/cors.go
@@ -1,4 +1,4 @@
-package main
+package handler
 
 import "net/http"
