+++ server/core/cmd/httpserver/handler/responsetype.go
@@ -1,4 +1,4 @@
-package main
+package handler
 
 import "net/http"
